+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "clam"
-version = "0.8.0"
+version = "0.9.0"
 authors = [
     "Najib Ishaq <najib_ishaq@zoho.com>",
     "Tom Howard <info@tomhoward.codes>",
@@ -33,52 +33,29 @@ debug = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bitvec = "1.0"
-num-traits = "0.2"
-rayon = "1.5"
-libm = "0.2"
-rand = "0.8"
+num-traits = "0.2.15"
+libm = "0.2.7"
+rand_chacha = "0.3.1"
+rand = "0.8.5"
+bitvec = "1.0.1"
+priority-queue = "1.3.1"
+rayon = "1.7.0"
 smartcore = "0.2.1"
 automl = "0.2.7"
-ndarray = "0.15.4"
-ndarray-npy = "0.8.1"
-priority-queue = "1.2"
-rand_chacha = "0.3"
-# easy-cast = "0.4.4"
-# eval-metrics = "1.0.1"
-# ordered-float = "3.0"
-# simplelog = "0.11.1"
-# structopt = "0.3.23"
-# sysinfo = "0.23.5"
-
-[dependencies.serde]
-version = "1.0"
-features = ["derive"]
 
 [dev-dependencies]
-float-cmp = "0.9"
-ndarray-npy = "0.8"
-log = "0.4"
+float-cmp = "0.9.0"
+ndarray-npy = "0.8.1"
+log = "0.4.18"
 env_logger = "0.10.0"
-rmp-serde = "1.1"
+rmp-serde = "1.1.1"
 num-format = "0.4.4"
 kdam = "0.3.0"
-
-[dev-dependencies.criterion]
-version = "0.5.1"
-features = ["html_reports"]
-
-[dev-dependencies.ndarray]
-version = "0.15"
-features = ["rayon"]
-
-[dev-dependencies.serde_json]
-version = "1.0"
-features = ["alloc"]
-
-[dev-dependencies.arrow]
-version = "40.0.0"
-features = ["ipc"]
+criterion = { version = "0.5.1", features = ["html_reports"] }
+serde = { version = "1.0.164", features = ["derive"] }
+serde_json = { version = "1.0.96", features = ["alloc"] }
+ndarray = { version = "0.15.6", features = ["rayon"] }
+arrow = { version = "41.0.0", features = ["ipc"] }
 
 [[example]]
 name = "reports"
