+++ benches/rnn-search.rs
@@ -1,7 +1,7 @@
 use criterion::*;
 
-use clam::core::cluster_criteria::PartitionCriteria;
-use clam::core::dataset::VecVec;
+use clam::cluster::PartitionCriteria;
+use clam::dataset::VecVec;
 use clam::distances::f32::METRICS;
 use clam::search::cakes::CAKES;
 use clam::utils::helpers;
