+++ src/core/cluster/mod.rs
@@ -1,0 +1,7 @@
+mod _cluster;
+mod criteria;
+mod tree;
+
+pub(crate) use _cluster::{Cluster, Ratios};
+pub use criteria::PartitionCriteria;
+pub use tree::Tree;
