+++ src/core/dataset/mod.rs
@@ -1,0 +1,5 @@
+mod _dataset;
+mod vec2d;
+
+pub use _dataset::Dataset;
+pub use vec2d::VecVec;
