+++ src/core/graph/_graph.rs
@@ -3,22 +3,23 @@ use std::collections::HashSet;
 use std::hash::Hash;
 use std::hash::Hasher;
 
-use super::cluster::Cluster;
-use super::dataset::Dataset;
-use super::number::Number;
+use crate::cluster::Cluster;
+use crate::dataset::Dataset;
+use crate::number::Number;
 
-type ClusterSet<'a, T, U, D> = HashSet<&'a Cluster<'a, T, U, D>>;
+type ClusterSet<'a, T, U, D> = HashSet<&'a Cluster<T, U, D>>;
 type EdgeSet<'a, T, U, D> = HashSet<&'a Edge<'a, T, U, D>>;
-type AdjacencyMap<'a, T, U, D> = HashMap<&'a Cluster<'a, T, U, D>, ClusterSet<'a, T, U, D>>;
-type FrontierSizes<'a, T, U, D> = HashMap<&'a Cluster<'a, T, U, D>, Vec<usize>>;
+type AdjacencyMap<'a, T, U, D> = HashMap<&'a Cluster<T, U, D>, ClusterSet<'a, T, U, D>>;
+type FrontierSizes<'a, T, U, D> = HashMap<&'a Cluster<T, U, D>, Vec<usize>>;
 
 /// Two `Cluster`s have an `Edge` between them if they have overlapping volumes.
 ///
 /// In CLAM, all `Edge`s are bi-directional.
 #[derive(Debug, Clone)]
-pub struct Edge<'a, T: Number, U: Number, D: Dataset<T, U>> {
-    left: &'a Cluster<'a, T, U, D>,
-    right: &'a Cluster<'a, T, U, D>,
+pub(crate) struct Edge<'a, T: Number, U: Number, D: Dataset<T, U>> {
+    left: &'a Cluster<T, U, D>,
+    right: &'a Cluster<T, U, D>,
+    #[allow(dead_code)]
     distance: U,
 }
 
@@ -49,7 +50,8 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Edge<'a, T, U, D> {
     ///
     /// It is upon the user to verify that the two `Cluster`s are close enough
     /// to have an edge between them.
-    pub fn new(left: &'a Cluster<'a, T, U, D>, right: &'a Cluster<'a, T, U, D>, distance: U) -> Self {
+    #[allow(dead_code)]
+    pub fn new(left: &'a Cluster<T, U, D>, right: &'a Cluster<T, U, D>, distance: U) -> Self {
         if left < right {
             Self { left, right, distance }
         } else {
@@ -62,31 +64,37 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Edge<'a, T, U, D> {
     }
 
     /// Whether this edge has the given `Cluster` at one of its ends.
-    pub fn contains(&self, c: &Cluster<'a, T, U, D>) -> bool {
+    #[allow(dead_code)]
+    pub fn contains(&self, c: &Cluster<T, U, D>) -> bool {
         c == self.left || c == self.right
     }
 
     /// A 2-slice of the `Cluster`s in this `Edge`.
-    pub fn clusters(&self) -> [&Cluster<'a, T, U, D>; 2] {
+    #[allow(dead_code)]
+    pub fn clusters(&self) -> [&Cluster<T, U, D>; 2] {
         [self.left, self.right]
     }
 
     /// A reference to the `Cluster` at the `left` end of the `Edge`.
-    pub fn left(&self) -> &Cluster<'a, T, U, D> {
+    #[allow(dead_code)]
+    pub fn left(&self) -> &Cluster<T, U, D> {
         self.left
     }
 
     /// A reference to the `Cluster` at the `right` end of the `Edge`.
-    pub fn right(&self) -> &Cluster<'a, T, U, D> {
+    #[allow(dead_code)]
+    pub fn right(&self) -> &Cluster<T, U, D> {
         self.right
     }
 
     /// The distance between the two `Cluster`s connected by this `Edge`.
+    #[allow(dead_code)]
     pub fn distance(&self) -> U {
         self.distance
     }
 
     /// Whether this is an edge from a `Cluster` to itself.
+    #[allow(dead_code)]
     pub fn is_circular(&self) -> bool {
         self.left == self.right
     }
@@ -96,7 +104,8 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Edge<'a, T, U, D> {
     /// Err:
     ///
     /// * If `c` is not one of the `Cluster`s connected by this `Edge`.
-    pub fn neighbor(&self, c: &Cluster<'a, T, U, D>) -> Result<&Cluster<'a, T, U, D>, String> {
+    #[allow(dead_code)]
+    pub fn neighbor(&self, c: &Cluster<T, U, D>) -> Result<&Cluster<T, U, D>, String> {
         if c == self.left {
             Ok(self.right)
         } else if c == self.right {
@@ -112,14 +121,14 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Edge<'a, T, U, D> {
 ///
 /// TODO: Add more info on what graphs we useful for.
 #[derive(Debug, Clone)]
-pub struct Graph<'a, T: Number, U: Number, D: Dataset<T, U>> {
+pub(crate) struct Graph<'a, T: Number, U: Number, D: Dataset<T, U>> {
     clusters: ClusterSet<'a, T, U, D>,
     edges: EdgeSet<'a, T, U, D>,
     adjacency_map: AdjacencyMap<'a, T, U, D>,
     population: usize,
     min_depth: usize,
     max_depth: usize,
-    ordered_clusters: Vec<&'a Cluster<'a, T, U, D>>,
+    ordered_clusters: Vec<&'a Cluster<T, U, D>>,
     distance_matrix: Option<Vec<Vec<U>>>,
     adjacency_matrix: Option<Vec<Vec<bool>>>,
     frontier_sizes: Option<FrontierSizes<'a, T, U, D>>, // TODO: Bench when replacing with DashMap
@@ -134,6 +143,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
     ///
     /// * `clusters`: The set of `Cluster`s with which to build the `Graph`.
     /// * `edges`: The set of `Edge`s with which to build the `Graph`.
+    #[allow(dead_code)]
     pub fn new(clusters: ClusterSet<'a, T, U, D>, edges: EdgeSet<'a, T, U, D>) -> Self {
         assert!(!clusters.is_empty());
 
@@ -171,6 +181,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
         }
     }
 
+    #[allow(dead_code)]
     fn compute_distance_matrix(&self) -> Vec<Vec<U>> {
         let indices: HashMap<_, _> = self.ordered_clusters.iter().enumerate().map(|(i, &c)| (c, i)).collect();
         let mut matrix: Vec<Vec<U>> = vec![vec![U::zero(); self.vertex_cardinality()]; self.vertex_cardinality()];
@@ -185,6 +196,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
 
     /// Computes the distance matrix for the `Graph` and stores it as an
     /// internal property.
+    #[allow(dead_code)]
     pub fn with_distance_matrix(mut self) -> Self {
         self.distance_matrix = Some(self.compute_distance_matrix());
         self
@@ -196,6 +208,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
     /// # Panics:
     ///
     /// * If called before calling `with_distance_matrix`.
+    #[allow(dead_code)]
     pub fn with_adjacency_matrix(mut self) -> Self {
         self.adjacency_matrix = Some(
             self.distance_matrix()
@@ -206,6 +219,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
         self
     }
 
+    #[allow(dead_code)]
     pub fn with_eccentricities(&'a self) -> Self {
         let frontier_sizes = Some(
             self.clusters
@@ -228,6 +242,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
         }
     }
 
+    #[allow(dead_code)]
     #[allow(clippy::manual_retain)]
     pub fn find_component_clusters(&'a self) -> Vec<ClusterSet<'a, T, U, D>> {
         let mut components = Vec::new();
@@ -247,58 +262,71 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
         components
     }
 
+    #[allow(dead_code)]
     pub fn clusters(&self) -> &ClusterSet<'a, T, U, D> {
         &self.clusters
     }
 
+    #[allow(dead_code)]
     pub fn edges(&self) -> &EdgeSet<'a, T, U, D> {
         &self.edges
     }
 
+    #[allow(dead_code)]
     pub fn vertex_cardinality(&self) -> usize {
         self.clusters.len()
     }
 
+    #[allow(dead_code)]
     pub fn edge_cardinality(&self) -> usize {
         self.edges.len()
     }
 
+    #[allow(dead_code)]
     pub fn population(&self) -> usize {
         self.population
     }
 
+    #[allow(dead_code)]
     pub fn min_depth(&self) -> usize {
         self.min_depth
     }
 
+    #[allow(dead_code)]
     pub fn max_depth(&self) -> usize {
         self.max_depth
     }
 
+    #[allow(dead_code)]
     pub fn depth_range(&self) -> (usize, usize) {
         (self.min_depth, self.max_depth)
     }
 
+    #[allow(dead_code)]
     pub fn adjacency_map(&'a self) -> &AdjacencyMap<T, U, D> {
         &self.adjacency_map
     }
 
-    pub fn ordered_clusters(&self) -> &[&Cluster<'a, T, U, D>] {
+    #[allow(dead_code)]
+    pub fn ordered_clusters(&self) -> &[&Cluster<T, U, D>] {
         &self.ordered_clusters
     }
 
+    #[allow(dead_code)]
     pub fn distance_matrix(&self) -> &[Vec<U>] {
         self.distance_matrix
             .as_ref()
             .expect("Please call `with_distance_matrix` on the Graph before using `distance_matrix`.")
     }
 
+    #[allow(dead_code)]
     pub fn adjacency_matrix(&self) -> &[Vec<bool>] {
         self.adjacency_matrix
             .as_ref()
             .expect("Please call `with_adjacency_matrix` on the Graph before using `adjacency_matrix`.")
     }
 
+    #[allow(dead_code)]
     pub fn diameter(&'a self) -> usize {
         self.clusters
             .iter()
@@ -307,7 +335,7 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
             .unwrap()
     }
 
-    fn assert_contains(&self, c: &Cluster<'a, T, U, D>) -> Result<(), String> {
+    fn assert_contains(&self, c: &Cluster<T, U, D>) -> Result<(), String> {
         if self.clusters.contains(&c) {
             Ok(())
         } else {
@@ -315,25 +343,28 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
         }
     }
 
-    pub fn unchecked_vertex_degree(&'a self, c: &Cluster<'a, T, U, D>) -> usize {
+    #[allow(dead_code)]
+    pub fn unchecked_vertex_degree(&'a self, c: &Cluster<T, U, D>) -> usize {
         self.unchecked_neighbors_of(c).len()
     }
 
-    pub fn vertex_degree(&'a self, c: &Cluster<'a, T, U, D>) -> Result<usize, String> {
+    #[allow(dead_code)]
+    pub fn vertex_degree(&'a self, c: &Cluster<T, U, D>) -> Result<usize, String> {
         self.assert_contains(c)?;
         Ok(self.unchecked_vertex_degree(c))
     }
 
-    pub fn unchecked_neighbors_of(&'a self, c: &Cluster<'a, T, U, D>) -> &ClusterSet<T, U, D> {
+    pub fn unchecked_neighbors_of(&'a self, c: &Cluster<T, U, D>) -> &ClusterSet<T, U, D> {
         self.adjacency_map.get(c).unwrap()
     }
 
-    pub fn neighbors_of(&'a self, c: &Cluster<'a, T, U, D>) -> Result<&ClusterSet<T, U, D>, String> {
+    #[allow(dead_code)]
+    pub fn neighbors_of(&'a self, c: &Cluster<T, U, D>) -> Result<&ClusterSet<T, U, D>, String> {
         self.assert_contains(c)?;
         Ok(self.unchecked_neighbors_of(c))
     }
 
-    pub fn unchecked_traverse(&'a self, start: &'a Cluster<'a, T, U, D>) -> (ClusterSet<T, U, D>, Vec<usize>) {
+    pub fn unchecked_traverse(&'a self, start: &'a Cluster<T, U, D>) -> (ClusterSet<T, U, D>, Vec<usize>) {
         let mut visited: HashSet<&Cluster<T, U, D>> = HashSet::new();
         let mut frontier: HashSet<&Cluster<T, U, D>> = HashSet::new();
         frontier.insert(start);
@@ -353,13 +384,15 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
         (visited, frontier_sizes)
     }
 
+    #[allow(dead_code)]
     #[allow(clippy::type_complexity)]
-    pub fn traverse(&'a self, start: &'a Cluster<'a, T, U, D>) -> Result<(ClusterSet<T, U, D>, Vec<usize>), String> {
+    pub fn traverse(&'a self, start: &'a Cluster<T, U, D>) -> Result<(ClusterSet<T, U, D>, Vec<usize>), String> {
         self.assert_contains(start)?;
         Ok(self.unchecked_traverse(start))
     }
 
-    pub fn unchecked_frontier_sizes(&'a self, c: &'a Cluster<'a, T, U, D>) -> &[usize] {
+    #[allow(dead_code)]
+    pub fn unchecked_frontier_sizes(&'a self, c: &'a Cluster<T, U, D>) -> &[usize] {
         self.frontier_sizes
             .as_ref()
             .expect("Please call `with_eccentricities` before using this method.")
@@ -367,16 +400,19 @@ impl<'a, T: Number, U: Number, D: Dataset<T, U>> Graph<'a, T, U, D> {
             .unwrap()
     }
 
-    pub fn frontier_sizes(&'a self, c: &'a Cluster<'a, T, U, D>) -> Result<&[usize], String> {
+    #[allow(dead_code)]
+    pub fn frontier_sizes(&'a self, c: &'a Cluster<T, U, D>) -> Result<&[usize], String> {
         self.assert_contains(c)?;
         Ok(self.unchecked_frontier_sizes(c))
     }
 
-    pub fn unchecked_eccentricity(&'a self, c: &'a Cluster<'a, T, U, D>) -> usize {
+    #[allow(dead_code)]
+    pub fn unchecked_eccentricity(&'a self, c: &'a Cluster<T, U, D>) -> usize {
         self.unchecked_frontier_sizes(c).len()
     }
 
-    pub fn eccentricity(&'a self, c: &'a Cluster<'a, T, U, D>) -> Result<usize, String> {
+    #[allow(dead_code)]
+    pub fn eccentricity(&'a self, c: &'a Cluster<T, U, D>) -> Result<usize, String> {
         self.assert_contains(c)?;
         Ok(self.unchecked_eccentricity(c))
     }
