+++ src/core/graph/criteria.rs
@@ -1,4 +1,4 @@
 //! Criteria used for selecting `Cluster`s for `Graph`s.
 
 /// A `Box`ed function that assigns a score for a given `Cluster`.
-pub type MetaMLScorer = Box<dyn (Fn(super::cluster::Ratios) -> f64) + Send + Sync>;
+pub type MetaMLScorer = fn(crate::cluster::Ratios) -> f64;
