+++ src/core/graph/mod.rs
@@ -1,0 +1,6 @@
+mod _graph;
+mod criteria;
+
+#[allow(unused_imports)]
+pub(crate) use _graph::{Edge, Graph};
+pub use criteria::MetaMLScorer;
