+++ src/core/mod.rs
@@ -1,9 +1,4 @@
 pub mod cluster;
-pub mod cluster_criteria;
 pub mod dataset;
-//The two modules below leak `Cluster` and need some revision.
-//pub mod graph;
-//pub mod graph_criteria;
-pub mod instance;
-pub mod manifold;
+pub mod graph;
 pub mod number;
