+++ src/core/number/mod.rs
@@ -1,0 +1,7 @@
+mod _number;
+mod num_bool;
+mod num_variants;
+
+pub use _number::Number;
+pub use num_bool::NumBool;
+pub use num_variants::{FloatNumber, IIntNumber, IntNumber, UIntNumber};
