+++ src/distances/mod.rs
@@ -1,6 +1,6 @@
 pub mod f32;
-pub mod u32;
 pub mod lp_norms;
 pub mod misc;
 pub mod sets;
 pub mod strings;
+pub mod u32;
