+++ src/lib.rs
@@ -6,8 +6,10 @@
 //! - [CHAODA](https://arxiv.org/abs/2103.11774)
 //!
 
-//pub mod anomaly;
-pub mod core;
+// pub mod anomaly;
+mod core;
 pub mod distances;
 pub mod search;
 pub mod utils;
+
+pub use crate::core::{cluster, dataset, graph, number};
